Move chatlog-related functions to their own file.
